build: bump to python 3.14 with modern dependency stack.
- requires-python >=3.14; numpy 2.5, torch 2.13, sklearn 1.9, pandas 3.0,
  tiktoken 0.13; drop unused torchtext; declare formerly-informal book deps
  (torchmetrics, transformers, datasets, statsmodels, opencv); modernize
  serving extra (sqlalchemy 3.14-compatible)
- mypy 2.2 + pyright target 3.14; tracked package mypy-clean (177 files)
- fix numpy-2/psutil-7/torch-2.13 breakages in package, tests, and two
  book pages (scheduler verbose removal)
- exclude 3 pages from notebook execution pending content modernization
  (memory_internals 05/06, gpt 05_adder)

+++ omnivault/machine_learning/mixture/gmm.py
@@ -104,7 +104,7 @@ class GaussianMixtureModel(BaseEstimator):
 
             labels = kmeans.labels
             # take kmeans labels and use them to initialize the GMM
-            self.weights_ = np.bincount(labels) / self.num_samples
+            self.weights_ = np.bincount(labels.astype(int)) / self.num_samples
             # means is the centroids of the kmeans
             self.means_ = kmeans.centroids
             # covariances is the covariance of each cluster k
